% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{create_fips}
\alias{create_fips}
\title{Create FIPS code for different geographies}
\usage{
create_fips(data, geography)
}
\arguments{
\item{data}{A tibble that is the result of Population Estimates API call}

\item{geography}{Either "us", "state", or "county".}
}
\value{
Character vector of FIPS codes.
}
\description{
Creates the FIPS vector for the population data
}
\keyword{internal}
